requirements.txt
Package                Version
---------------------- -----------
asgiref                3.5.2
certifi                2022.9.24
cffi                   1.15.1
charset-normalizer     2.1.1
click                  8.1.3
colorama               0.4.6
contourpy              1.0.5
cryptography           38.0.1
cycler                 0.11.0
dataschema             0.1
Django                 4.1.2
django-filter          22.1
djangorestframework    3.14.0
et-xmlfile             1.1.0
Faker                  15.1.1
Flask                  2.2.2
fonttools              4.38.0
greenlet               1.1.3.post0
idna                   3.4
itsdangerous           2.1.2
Jinja2                 3.1.2
kiwisolver             1.4.4
Markdown               3.4.1
MarkupSafe             2.1.1
matplotlib             3.6.1
mysql-connector        2.2.9
mysql-connector-python 8.0.31
mysqlclient            2.1.1
numpy                  1.23.4
openpyxl               3.0.10
packaging              21.3
pandas                 1.5.1
pathlib                1.0.1
Pillow                 9.2.0
pip                    22.3
protobuf               3.20.1
pycparser              2.21
pyfarmhash             0.3.2
PyMySQL                1.0.2
pyparsing              3.0.9
python-dateutil        2.8.2
pytz                   2022.5
requests               2.28.1
setuptools             63.2.0
six                    1.16.0
SQLAlchemy             1.4.42
sqlparse               0.4.3
tzdata                 2022.6
urllib3                1.26.12
Werkzeug               2.2.2
xlrd                   2.0.1
